Add files via upload
Me and a partner developed an inventory management system for the Westminster Ski Team. The database tracks all the athletes skis, the brands as well as information about each skier. It has a interface that is connected to the database making it possible for the user to add, update or delete parts as they please.