Add files via upload
This project is a chat room written in Java. The programs lets the users instantly send and receive messages across the network. 

I chose this project cause I think it was the most fun project to get working. Building the program step by step, and finally being able to instantly send messages to your classmates was very rewarding. Even though this was a individual assignment, we had to design the protocol as a team. This taught me how to work in a group setting as we had to discuss the pros and cons of different solutions and then make a decision on a design the entire team would follow.